+++ custom_components/supernotify/delivery.py
@@ -43,6 +43,7 @@ _LOGGER = logging.getLogger(__name__)
 
 class Delivery(DeliveryConfig):
     def __init__(self, name: str, conf: ConfigType, transport: "Transport") -> None:
+        conf = conf or {}
         self.name: str = name
         self.alias: str | None = conf.get(CONF_ALIAS)
         self.transport: Transport = transport